refactor(viewer): extract PdfViewerControl into standalone Pdfe.Avalonia library (#365)
Move the embedded Avalonia PDF viewer out of the PdfEditor app into a new,
dependency-light control library `Pdfe.Avalonia` (depends only on Pdfe.Core +
Pdfe.Rendering + Avalonia + SkiaSharp), so any Avalonia app can drop in a
pure-managed, SkiaSharp-based PDF viewer. The app now consumes it as the
reference implementation; behavior is unchanged.

Moved (history preserved): PdfViewerControl.axaml(.cs), SkiaInterop (Imaging),
TextSelectionEngine (Services), and the HiddenTextHighlight overlay model —
all renamespaced to Pdfe.Avalonia.*. Event-arg types + InteractionMode are
public; the control's HiddenTextHighlights now binds the library-owned type, so
the app's VM collection binds directly (no projection layer).

Two namespace gotchas resolved: naming the library `Pdfe.Avalonia` shadows the
real `Avalonia` root namespace for fully-qualified inline refs (fixed with
`global::Avalonia.*`); and the `Subscribe(Action<T>)` extension came from
System.Reactive via the app's ReactiveUI — the library has no Rx dependency, so
the viewport subscription now uses Avalonia's `AnonymousObserver<Size>`.

App + tests updated to the new namespaces (xmlns, usings). Full solution builds
0 warnings; 52 control/UI/engine tests pass.
