+++ PdfEditor.Tests/UI/FormFieldsOverlayTests.cs
@@ -9,7 +9,7 @@ using Avalonia.Interactivity;
 using Avalonia.Headless.XUnit;
 using AwesomeAssertions;
 using Pdfe.Core.Document;
-using PdfEditor.Controls;
+using Pdfe.Avalonia.Controls;
 using PdfEditor.ViewModels;
 using PdfEditor.Views;
 using Xunit;
