+++ PdfEditor.Tests/UI/TextSelectionDragTests.cs
@@ -13,7 +13,8 @@ using Avalonia.Threading;
 using AwesomeAssertions;
 using Pdfe.Core.Document;
 using Pdfe.Core.Text;
-using PdfEditor.Controls;
+using Pdfe.Avalonia.Controls;
+using Pdfe.Avalonia.Services;
 using PdfEditor.Services;
 using PdfEditor.ViewModels;
 using PdfEditor.Views;
