+++ PdfEditor/Views/MainWindow.axaml
@@ -2,7 +2,7 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:vm="using:PdfEditor.ViewModels"
         xmlns:models="using:PdfEditor.Models"
-        xmlns:controls="using:PdfEditor.Controls"
+        xmlns:controls="using:Pdfe.Avalonia.Controls"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:ui="using:FluentAvalonia.UI.Controls"
