+++ Pdfe.Avalonia/Controls/PdfViewerControl.axaml
@@ -3,7 +3,7 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="600"
-             x:Class="PdfEditor.Controls.PdfViewerControl">
+             x:Class="Pdfe.Avalonia.Controls.PdfViewerControl">
 
     <Grid>
         <!-- Main PDF display area.
